Create a /BackButton folder in /client/components and move the BackButton files there
