+++ src/client/pages/Checkout/shared/CheckoutPageWrapper.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import styled from '@emotion/styled'
 import { colorsV3 } from '@hedviginsurance/brand'
 import { TopBar } from 'components/TopBar'
-import { BackButton } from './BackButton'
+import { BackButton } from 'components/BackButton/BackButton'
 
 const { gray100 } = colorsV3
